Revert use of PTY in windows
